requirements.txt
Flask
Flask-Cors
Authlib
PyJWT
gunicorn
pandas
python-dateutil
Jinja2
requests
psycopg2-binary
